Dev seed pulls in Map Roulette challenge users to database. These mostly do not exist in OSMesa
